rename to code_generators: templates are just a means
